Refactor du code + Modification fonction getLink pour retourner le lien (et non le hash) + ajout du hash en tant que clé primaire
